src/requirements.txt
Django==4.1.7
django_ckeditor==6.5.1
django_embed_video==1.4.8
django_filter==22.1
django_mptt==0.14.0
Pillow==10.0.0
xlwt==1.3.0
